requirements.txt
astroid==2.12.10
Babel==2.10.3
beautifulsoup4==4.11.1
black==22.8.0
bs4==0.0.1
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
dill==0.3.5.1
fonttools==4.37.3
idna==3.4
isort==5.10.1
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
matplotlib==3.6.0
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.23.3
packaging==21.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pylint==2.15.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tk==0.1.0
tkcalendar==1.6.1
tomli==2.0.1
tomlkit==0.11.4
typing-extensions==4.3.0
urllib3==1.26.12
wrapt==1.14.1
xmltodict==0.13.0
